+++ tests/Application/Command/VideoConference/RequestAccessHandlerTest.php
@@ -1,16 +1,12 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
+namespace Proximum\Vimeet\Tests\Application\Command\VideoConference;
 
 use OpenTok\Session;
 use PHPUnit\Framework\TestCase;
 use Proximum\Vimeet\Application\Adapter\VideoConferenceAdapterInterface;
+use Proximum\Vimeet\Application\Command\Meeting\AddParticipantToMeeting;
+use Proximum\Vimeet\Application\Command\Meeting\AddParticipantToMeetingHandler;
 use Proximum\Vimeet\Application\Command\VideoConference\RequestAccess;
 use Proximum\Vimeet\Application\Command\VideoConference\RequestAccessHandler;
 use Proximum\Vimeet\Application\Components\Visio\VisioSettingsRetriever;
@@ -18,6 +14,8 @@ use Proximum\Vimeet\Application\View\Meeting\VideoConferenceView;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Meeting;
 use Proximum\Vimeet\Domain\Model\MeetingSlot;
+use Proximum\Vimeet\Domain\Model\Participant;
+use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\User;
 use Proximum\Vimeet\Domain\Model\VideoConference;
 use Proximum\Vimeet\Domain\Model\VideoConferenceToken;
@@ -32,6 +30,11 @@ class RequestAccessHandlerTest extends TestCase
         $meeting = $this->prophesize(Meeting::class);
         $slot    = $this->prophesize(MeetingSlot::class);
         $user    = $this->prophesize(User::class);
+
+        $sheet = $this->prophesize(Sheet::class);
+        $participant = $this->prophesize(Participant::class);
+        $participant->getSheet()->shouldBeCalled()->willReturn($sheet->reveal());
+        $participant->getUser()->shouldBeCalled()->willReturn($user->reveal());
 
         $slot->getEnd()->shouldBeCalled()->willReturn($slotEnd);
         $meeting->getSlot()->shouldBeCalled()->willReturn($slot->reveal());
@@ -74,7 +77,14 @@ class RequestAccessHandlerTest extends TestCase
 
         $videoConferenceRepository->add($videoConference)->shouldBeCalled();
 
+        $addParticipantToMeetingHandler = $this->prophesize(AddParticipantToMeetingHandler::class);
+        $addParticipantToMeetingHandler
+            ->handle(new AddParticipantToMeeting($participant->reveal(), $meeting->reveal()))
+            ->shouldBeCalled()
+        ;
+
         $handler = new RequestAccessHandler(
+            $addParticipantToMeetingHandler->reveal(),
             $videoConferenceAdapter->reveal(),
             $videoConferenceRepository->reveal(),
             $visioSettingsRetriever->reveal()
@@ -94,7 +104,7 @@ class RequestAccessHandlerTest extends TestCase
         $videoConferenceView = $handler->handle(
             new RequestAccess(
                 $meeting->reveal(),
-                $user->reveal(),
+                $participant->reveal(),
                 'fr'
             )
         );
@@ -108,6 +118,11 @@ class RequestAccessHandlerTest extends TestCase
         $meeting = $this->prophesize(Meeting::class);
         $slot    = $this->prophesize(MeetingSlot::class);
         $user    = $this->prophesize(User::class);
+
+        $sheet = $this->prophesize(Sheet::class);
+        $participant = $this->prophesize(Participant::class);
+        $participant->getSheet()->shouldBeCalled()->willReturn($sheet->reveal());
+        $participant->getUser()->shouldBeCalled()->willReturn($user->reveal());
 
         $slot->getEnd()->shouldBeCalled()->willReturn($slotEnd);
         $meeting->getSlot()->shouldBeCalled()->willReturn($slot->reveal());
@@ -158,7 +173,14 @@ class RequestAccessHandlerTest extends TestCase
 
         $videoConferenceRepository->set($videoConference)->shouldBeCalled();
 
+        $addParticipantToMeetingHandler = $this->prophesize(AddParticipantToMeetingHandler::class);
+        $addParticipantToMeetingHandler
+            ->handle(new AddParticipantToMeeting($participant->reveal(), $meeting->reveal()))
+            ->shouldBeCalled()
+        ;
+
         $handler = new RequestAccessHandler(
+            $addParticipantToMeetingHandler->reveal(),
             $videoConferenceInterface->reveal(),
             $videoConferenceRepository->reveal(),
             $visioSettingsRetriever->reveal()
@@ -178,7 +200,7 @@ class RequestAccessHandlerTest extends TestCase
         $videoConferenceView = $handler->handle(
             new RequestAccess(
                 $meeting->reveal(),
-                $user->reveal(),
+                $participant->reveal(),
                 'fr'
             )
         );